### 2018-11-28 #1105 @whedon
[PRE REVIEW]: The SuiteSparse Matrix Collection Website Interface
**Submitting author:** @ScottKolo (<a href="http://orcid.org/0000-0001-8340-0824">Scott Kolodziej</a>)
**Repository:** <a href="https://github.com/ScottKolo/suitesparse-matrix-collection-website" target ="_blank">https://github.com/ScottKolo/suitesparse-matrix-collection-website</a>
**Version:** JOSS-1.0.0
