### 2018-12-27 4bc28e8
Fix JOSS paper bibliography formatting issues
