+++ paper/bibliography.bib
@@ -1,6 +1,6 @@
 @article{SSCollection,
  author = {Davis, Timothy A. and Hu, Yifan},
- title = {The University of Florida Sparse Matrix Collection},
+ title = {{T}he {U}niversity of {F}lorida {S}parse {M}atrix {C}ollection},
  journal = {ACM Trans. Math. Softw.},
  issue_date = {November 2011},
  volume = {38},
@@ -29,14 +29,14 @@
 }
 
 @techreport{RutherfordBoeing,
-  title={The Rutherford-Boeing sparse matrix collection},
+  title={{T}he {R}utherford-{B}oeing sparse matrix collection},
   author={Duff, Iain S and Grimes, Roger G and Lewis, John G},
   year={1997},
   institution={Council for the Central Laboratory of the Research Councils}
 }
 
 @misc{ClarivateDCI,
-  title={Data Citation Index - Clarivate Analytics},
+  title={{D}ata {C}itation {I}ndex - {C}larivate {A}nalytics},
   publisher={Clarivate Analytics},
   url={http://wokinfo.com/products_tools/multidisciplinary/dci/}
 }
@@ -48,7 +48,7 @@
 }
 
 @misc{re3dataSuiteSparse,
-  title={re3data.org: The SuiteSparse Matrix Collection}, 
+  title={re3data.org: {T}he {S}uite{S}parse {M}atrix {C}ollection}, 
   organization={re3data.org - Registry of Research Data Repositories},
   url={http://doi.org/10.17616/R3WG6Z},
   doi={10.17616/R3WG6Z},
@@ -74,8 +74,8 @@
 }
 
 @article{davisalgorithm,
-  title={Algorithm XXX: Mongoose, A Graph Coarsening and Partitioning Library},
-  author={Davis, Timothy A And Hager, William W And Kolodziej, Scott P And Yeralan, S Nuri},
+  title={Algorithm {XXX}: Mongoose, A Graph Coarsening and Partitioning Library},
+  author={Davis, Timothy A., and Hager, William W., and Kolodziej, Scott P., and Yeralan, S Nuri},
   howpublished={Submitted}
 }
 
