### 2018-12-27 #1105 @ScottKolo
Hello, I just made a few tweaks to the paper bibliography to fix some formatting issues and had the PDF regenerated to reflect the changes.

Can I ask about the status of our submission? Do you need anything further on our end? I know the holidays kind of slow things down, I just wanted to see if we are still in pre-review.

Thank you!
