### 2018-12-28 #1105 @jedbrown
I'm waiting to hear back from potential reviewers. This submission is a bit unusual for JOSS in that the code is not itself research (though it represents significant effort and is important for facilitating research), but I think it's still fair to review according to the usual criteria. Feel free to make suggestions for potential reviewers.
