### 2019-02-08 #1244 @whedon
[REVIEW]: The SuiteSparse Matrix Collection Website Interface
**Submitting author:** @ScottKolo (<a href="http://orcid.org/0000-0001-8340-0824">Scott Kolodziej</a>)
**Repository:** <a href="https://github.com/ScottKolo/suitesparse-matrix-collection-website" target ="_blank">https://github.com/ScottKolo/suitesparse-matrix-collection-website</a>
**Version:** JOSS-1.0.2
**Archive:** <a href="https://doi.org/10.6084/m9.figshare.7381076" target="_blank">10.6084/m9.figshare.7381076</a>

- [x] **Version:** JOSS-1.0.2

- [x] **Version:** JOSS-1.0.2
