### 2019-02-08 #1244 @jedbrown
@dblarremore @tkonolige :wave: Welcome and thanks for agreeing to review! The comments from @whedon above outline the review process, which takes place in this thread (possibly with issues filed in the project repository). I'll be watching this thread if you have any questions.
