### 2019-02-11 ScottKolo/suitesparse-matrix-collection-website#5 @tkonolige
Documentation on building and testing
There is no documentation on how to build, run, and test the website locally. It would be nice to have for contributors.
