### 2019-02-11 #1244 @tkonolige
@ScottKolo I think https://doi.org/10.1007/978-1-5041-2940-4_9 is the correct doi for Matrix market: A web resource for test matrix collections. It would be great if you can add that.

I also see that there is no documentation for building, installing, and testing locally. I created an issue for it in your repository: ScottKolo/suitesparse-matrix-collection-website#5.
