+++ paper/bibliography.bib
@@ -25,7 +25,9 @@
   booktitle={Quality of Numerical Software},
   pages={125--137},
   year={1997},
-  publisher={Springer}
+  publisher={Springer},
+  url={https://doi.org/10.1007/978-1-5041-2940-4_9},
+  doi={10.1007/978-1-5041-2940-4_9}
 }
 
 @techreport{RutherfordBoeing,
