### 2019-02-12 442c76c
Add information about building and testing locally to README
