### 2019-02-12 ScottKolo/suitesparse-matrix-collection-website#5 @ScottKolo
Thanks for raising this issue @tkonolige. I've added information about building, running, and testing the web application locally to the main README.md file. 442c76cbabfb5d94ab5bdf7b26ab5f111c4c21d0
