### 2019-02-12 #1244 @tkonolige
@jedbrown I'm not sure about the example usage and functional documentation parts of the checklist. Does there need to be documentation on how to use the website? I think it is pretty straight forward.
