### 2019-02-13 8f5ff3f
Add features and functionality, URL routing information, and other interfaces section to README
