### 2019-02-13 aa001af
Unescape < and > in README (I underestimated GitHub).
