### 2019-02-13 e47ccec
Correctly escape angle brackets in GitHub Flavored Markdown
