### 2019-02-13 #1244 @ScottKolo
> I don't think human "use" of the website needs separate documentation, but there appears to be an API used by several external packages and I don't immediately notice documentation of that API.

We do include information about other interfaces to the Collection on the website itself on the [Interfaces page](https://sparse.tamu.edu/interfaces). I also just added information in the repository README referencing the Interfaces page and also the URL routing feature, which might be considered an API.

The use of most of the website should be (*we hope*) easy enough to understand and figure out. We also include tooltips to help users with certain terminology:

![image](https://user-images.githubusercontent.com/3859035/52680960-dbc26880-2eff-11e9-8d7f-f72e38e803dd.png)

Regarding the inner workings of things like UFget/ssget, we considered that somewhat out of scope for this paper as these interfaces were presented in [Davis and Hu](https://doi.org/10.1145/2049662.2049663) and are not strictly considered part of the web application. For example, these interfaces often provide the option to download a matrix, and that download URL is the same URL being used by the web application when you click the button on the website. The files are also not served by the Rails web application we've presented - they're being served by a file server. The way we see it, this is a "frontend" to access the "backend" presented in [Davis and Hu](https://doi.org/10.1145/2049662.2049663).
