### 2019-02-13 #1244 @tkonolige
That all makes sense. I think you're good on the documentation front then.

@ScottKolo, @jedbrown should the release version listed here be bumped to match what is currently in the repository (now that some changes have been made)?
