+++ db/collection_data/matrices/Martin/marine1.rb
@@ -1,0 +1,60 @@
+{
+    matrix_id: '2849',
+    name: 'marine1',
+    group: 'Martin',
+    description: 'chemical oceanography; a marine nitrogen cycle inverse model',
+    author: 'T. Martin, F. Primeau, K. L. Casciotti',
+    editor: 'T. Davis',
+    date: '2018',
+    kind: 'chemical oceanography problem',
+    problem_2D_or_3D: '0',
+    num_rows: '400320',
+    num_cols: '400320',
+    nonzeros: '6226538',
+    num_explicit_zeros: '0',
+    num_strongly_connected_components: '1',
+    pattern_symmetry: '0.922',
+    numeric_symmetry: '0.000',
+    rb_type: 'real',
+    structure: 'unsymmetric',
+    cholesky_candidate: 'no',
+    positive_definite: 'no',
+    notes: 'Martin/marine1: chemical oceanography; a marine nitrogen cycle inverse model   
+                                                                               
+A matrix submitted by Taylor Martin, Stanford, discussed the following paper:  
+                                                                               
+Title: Modeling oceanic nitrite concentrations and isotopes using a 3D         
+inverse N cycle model                                                          
+Authors: Taylor S. Martin(1), Francois Primeau(2), and Karen L. Casciotti(1)   
+(1) Stanford University, Department of Earth System Science                    
+(2) University of California, Irvine, Department of Earth System Science       
+Received: 05 Sep 2018                                                          
+Abstract. Nitrite (NO2-) is a key intermediate in the marine nitrogen (N) cycle
+and a substrate in nitrification, which produces nitrate (NO3-), as well as    
+water column N loss processes, denitrification and anammox. In models of the   
+marine N cycle, NO2- is often not considered as a separate state variable,     
+since NO3- occurs in much higher concentrations in the ocean. In oxygen        
+deficient zones (ODZs), however, NO2- represents a substantial fraction of the 
+bioavailable N, and modeling its production and consumption is important to    
+understanding the N cycle processes occurring there, especially those where    
+bioavailable N is lost from or retained within the water column. Here we       
+present the expansion of a global 3D inverse N cycle model to include NO2- as a
+reactive intermediate as well as the processes that produce and consume NO2- in
+marine ODZs. NO2- accumulation in ODZs is accurately represented by the model  
+involving NO3- reduction, NO2- reduction, NO2- oxidation, and anammox. We model
+both 14N and 15N and use a compilation of oceanographic measurements of NO3-   
+and NO2- concentrations and isotopes to place a better constraint on the N     
+cycle processes occurring. The model is optimized using a range of isotope     
+effects for denitrification and NO2- oxidation, and we find that the larger    
+(more negative) inverse isotope effects for NO2- oxidation along with          
+relatively high rates of NO2- oxidation give a better simulation of NO3- and   
+NO2- concentrations and isotopes in marine ODZs.                               
+                                                                               
+How to cite: Martin, T. S., Primeau, F., and Casciotti, K. L.: Modeling        
+oceanic nitrite concentrations and isotopes using a 3D inverse N cycle model,  
+Biogeosciences Discuss., https://doi.org/10.5194/bg-2018-397, in review, 2018. 
+',
+    b_field: 'full 400320-by-1
+',
+    image_files: 'marine1.png,',
+}
