+++ README.md
@@ -1,6 +1,6 @@
 # SuiteSparse Matrix Collection
 
-[![Build Status](https://semaphoreci.com/api/v1/scottkolo/suitesparse-matrix-collection-website/branches/master/badge.svg)](https://semaphoreci.com/scottkolo/suitesparse-matrix-collection-website) [![Code Climate](https://codeclimate.com/github/ScottKolo/suitesparse-matrix-collection-website/badges/gpa.svg)](https://codeclimate.com/github/ScottKolo/suitesparse-matrix-collection-website) [![Test Coverage](https://codeclimate.com/github/ScottKolo/suitesparse-matrix-collection-website/badges/coverage.svg)](https://codeclimate.com/github/ScottKolo/suitesparse-matrix-collection-website/coverage) [![Issue Count](https://codeclimate.com/github/ScottKolo/suitesparse-matrix-collection-website/badges/issue_count.svg)](https://codeclimate.com/github/ScottKolo/suitesparse-matrix-collection-website)
+[![Build Status](https://semaphoreci.com/api/v1/scottkolo/suitesparse-matrix-collection-website/branches/master/badge.svg)](https://semaphoreci.com/scottkolo/suitesparse-matrix-collection-website) [![Code Climate](https://codeclimate.com/github/ScottKolo/suitesparse-matrix-collection-website/badges/gpa.svg)](https://codeclimate.com/github/ScottKolo/suitesparse-matrix-collection-website) [![Test Coverage](https://codeclimate.com/github/ScottKolo/suitesparse-matrix-collection-website/badges/coverage.svg)](https://codeclimate.com/github/ScottKolo/suitesparse-matrix-collection-website/coverage) [![Issue Count](https://codeclimate.com/github/ScottKolo/suitesparse-matrix-collection-website/badges/issue_count.svg)](https://codeclimate.com/github/ScottKolo/suitesparse-matrix-collection-website) [![status](http://joss.theoj.org/papers/965b1c7bba801c85e9341a3db9013ef8/status.svg)](http://joss.theoj.org/papers/965b1c7bba801c85e9341a3db9013ef8)
 
 A web interface for the SuiteSparse Matrix Collection, formerly known as the University of Florida Sparse Matrix Collection. It is currently deployed at [https://sparse.tamu.edu](https://sparse.tamu.edu).
 
