### 2019-02-20 #1244 @dblarremore
**Trivial**: "applicaiton" typo in README.

**Perhaps more useful**:   Possible change needed in installation instructions for OSX.
I tried to do the local installation, updated Ruby, and got all of that sorted out. On README line
`bundle exec rake db:setup`
I kept getting an error
`PG::ConnectionBad: could not connect to server: No such file or directory`
`	Is the server running locally and accepting`
`	connections on Unix domain socket "/tmp/.s.PGSQL.5432"?`
I eventually fixed this by (apparently) manually starting the server via the top answer [here](https://dba.stackexchange.com/questions/75214/postgresql-not-running-on-mac) whose details are then described in detail [here](https://stackoverflow.com/questions/7975556/how-to-start-postgresql-server-on-mac-os-x).
After that, my local installation was totally fine.

To recap, my experience required that I insert this line between the bundle install and bundle exec rake commands:
`pg_ctl -D /usr/local/var/postgres -l /usr/local/var/postgres/server.log start`

Now, the question: is this worth inserting into the documentation? Are there files that could be modified by @ScottKolo or is this a sort of novice ruby user thing? (I am basically a test case for _user who does not know what he is doing with a local ruby installation_ so perhaps not an issue.) Curious to hear what you guys think.
