### 2019-02-20 1b8da80
Fix "applicaiton" typo in README
