### 2019-02-20 #1244 @ScottKolo
> To recap, my experience required that I insert this line between the bundle install and bundle exec rake commands:
`pg_ctl -D /usr/local/var/postgres -l /usr/local/var/postgres/server.log start`

After installing PostgreSQL, one needs to start it up so it's running in the background. You can do this a number of ways: manually start/stop, use the daemon, etc. After installing PostgreSQL from a package manager, it often spits out a line at the end about this (i.e. "run this magical incantation to start the daemon").

I'm happy to add whatever you think is helpful to the documentation. I might suggest the following additional call to `pg_isready`:

```shell
ruby -v                              # Check that Ruby 2.5+ is installed
pg_isready                           # Check that PostgreSQL database is ready
git clone https://github.com/ScottKolo/suitesparse-matrix-collection-website.git
cd suitesparse-matrix-collection-website
gem install bundler:1.16.6           # Dependency manager
bundle install --without production  # Download and install dependencies
bundle exec rake db:setup            # Generate and seed the database
bundle exec rails server             # Start the Ruby on Rails server application
```

Unfortunately, if `pg_isready` fails, it is not very informative regarding how to fix the problem.

I can add a line mentioning something about "PostgreSQL must be running - if it is not, try running `pg_ctl -D /usr/local/var/postgres -l /usr/local/var/postgres/server.log start`."

My only concern is that the above command assumes the user's `postgres` installation directory and log file location.
