### 2019-02-20 #1244 @dblarremore
I like the idea to use `pg_isready` but how about putting this snippet above, under the bullet list that includes the need for ruby and postgreSQL?  Then it would be filed in with the prep work and dependencies instead of with the installation itself.

What you wrote was really nice, so I tinkered with it a little, but maybe something like this? Happy for any modifications to this, since it's your README, but thought this might be useful:

> After installing PostgreSQL, note that one needs to start it up so it's running in the background. The status can be checked using `pg_isready`. If this fails, the installation can be started manually, e.g. via `pg_ctl -D /usr/local/var/postgres -l /usr/local/var/postgres/server.log start`, depending on the location of your installation. If PostgreSQL was installed from a package manager, it often spits out a line at the end about this (i.e. "run this magical incantation to start the daemon").

If you think this is an ok suggestion, and decide to put the `pg_isready` note above in the dependencies, it might also be appropriate to move the `ruby -v` check up there as well.
