### 2019-02-21 384f0ad
Add additional information regarding PostgreSQL and Ruby dependencies
