### 2019-02-21 #1244 @ScottKolo
With a little of my own wordsmithing, I have added information about starting PostgreSQL and installing Ruby to the README. Let me know if you think it's too much or not enough.

scottkolo/suitesparse-matrix-collection-website@384f0ad354c2d5be1b0351bb3aec5c9a0adc0955
