### 2019-03-08 c2a7540
JOSS paper: attempt sensible APA-like citations
