+++ paper/paper.md
@@ -45,7 +45,7 @@ bibliography: bibliography.bib
 
 The SuiteSparse Matrix Collection (formerly known as the University of Florida Sparse Matrix Collection) [@SSCollection] has grown significantly since its introduction, with newly added matrices representing almost seven times as much data than the entirety of the original Collection. With this growth, searching the Collection for matrices with specific names, structures, and other properties has become increasingly difficult. To make the Collection more accessible to the scientific computing community, we have developed a web application that allows real-time search and filtering of the Collection matrices.
 
-Built on Ruby on Rails, the web application was developed with software engineering best practices such as test-driven development, continuous integration via Semaphore CI, and static analysis and code coverage via Code Climate. Collectively, the web application and data storage now serve as the canonical source of the Collection from which other services, including the Clarivate Data Citation Index [@ClarivateDCI] and re3data [@re3dataSuiteSparse] reference or mirror the Collection. A variety of interfaces for accessing the Collection, including `ssget` in MATLAB, also obtain their data from this application.
+Built on Ruby on Rails, the web application was developed with software engineering best practices such as test-driven development, continuous integration via Semaphore CI, and static analysis and code coverage via Code Climate. Collectively, the web application and data storage now serve as the canonical source of the Collection from which other services, including the Clarivate Data Citation Index [-@ClarivateDCI] and re3data [-@re3dataSuiteSparse] reference or mirror the Collection. A variety of interfaces for accessing the Collection, including `ssget` in MATLAB, also obtain their data from this application.
 
 The SuiteSparse Matrix Collection has become the *lingua franca* of sparse matrix data and benchmarking, but its original website was written in static HTML, prohibiting any real-time searching or filtering of the Collection. With the current size, breadth, and variety of users of the Collection, this new web application provides to the scientific computing community a level of accessibility to the Collection not available before.
 
