### 2019-03-08 #1244 @jedbrown
Whedon is quite flaky about checking DOIs; they look fine in the paper. I made a PR for some bib formatting issues: https://github.com/ScottKolo/suitesparse-matrix-collection-website/pull/6
I think we'll be ready to archive and accept if that looks reasonable. Thanks for your patience.
