### 2019-03-08 #1244 @arfon
> Whedon is quite flaky about checking DOIs

I think Whedon only checks DOIs for `@article` entries currently. I'll see if I can fix this.
