### 2019-03-08 #1244 @ScottKolo
I have checked all of the suggested DOIs - they are all incorrect. Here's a quick discussion about each of these references and why they (currently) lack DOIs:

>MISSING DOIs
>
>- https://doi.org/10.1145/1057588.1057590 may be missing for title: The Rutherford-Boeing sparse matrix collection

This is a tech report. The suggested DOI references a paper that predates it by 15 years.

>- https://doi.org/10.14744/anatoljcardiol.2017.7846 may be missing for title: Data Citation Index - Clarivate Analytics

This is a website for Clarivate Analytics' Data Citation Index. The suggested DOI leads to a cardiology journal (?).

>- https://doi.org/10.1007/978-1-4842-2589-9_10 may be missing for title: University of Florida Sparse Matrix Collection - AWS Public Data Set

This is a website describing the use of the UFL Sparse Matrix Collection via AWS. The suggested DOI points to a generic article on Database Services in AWS.

>- https://doi.org/10.1016/s0092-8240(05)80013-8 may be missing for title: Efficient, high-quality force-directed graph drawing

This article was published in the Mathematica Journal, but it does not seem to have any DOI associated with it. The suggested DOI is a broken link.

>- https://doi.org/10.1016/j.jpdc.2015.06.005 may be missing for title: An improved exact algorithm and an NP-completeness proof for sparse matrix bipartitioning

This article is currently a preprint on arXiv. The suggested DOI links to some prior work by the same authors.

>- https://doi.org/10.1007/978-1-4757-3145-3_17 may be missing for title: Algorithm XXX: Mongoose, A Graph Coarsening and Partitioning Library

This article is currently in review. The suggested DOI references work by similar authors almost 20 years ago.

>- https://doi.org/10.2172/1128467 may be missing for title: A Projection Method for Metric-Constrained Optimization

This article is a preprint on arXiv. The suggested DOI points to an unrelated article in physics.
