### 2019-03-08 #1244 @jedbrown
Yeah, the DOI checker is very buggy. References look good to me now. Can you fix one spelling error "constrainted", tag a release ([annotated tag](https://git-scm.com/book/en/v2/Git-Basics-Tagging#_annotated_tags) preferred), and archive on Zenodo or similar?
