### 2019-03-08 #1244 @ScottKolo
One last nitpick: Is there any way to prevent line breaks in the middle of URLs?

![image](https://user-images.githubusercontent.com/3859035/54043531-75dd9f80-4192-11e9-8617-2b6c4ea26bab.png)
