### 2019-03-08 #1244 @jedbrown
I would consider dropping the `https://` part because it's arguably clutter given that full links are embedded. and/or using a tt font:
```
[`https://sparse.tamu.edu/Mycielski`](https://sparse.tamu.edu/Mycielski)
```
