### 2019-03-08 44d66e0
Simplify URLs and change to tt font to avoid mid-URL line breaks
