### 2019-03-08 cb5aecd
Remove tt font from URLs
