### 2019-03-08 #1244 @ScottKolo
Okay, I have tagged release at JOSS-1.0.2.
https://github.com/ScottKolo/suitesparse-matrix-collection-website/releases/tag/JOSS-1.0.2

We have archived the code for this release using figshare: https://figshare.com/articles/SuiteSparse_Matrix_Collection_Website/7381076

Does all of that look correct?
