### 2019-03-09 #1244 @jedbrown
Author list on the figshare appears to be missing Yifan Hu, who was on the paper. Can you fix the author lists to match in content and order?
