### 2019-03-09 #1244 @ScottKolo
Fixed. The Figshare DOI is https://doi.org/10.6084/m9.figshare.7381076

The title now matches, as do the authors and author order.
