### 2019-03-09 #1244 @jedbrown
@ScottKolo Is it intended that `https://` appears in exactly one of the URLs?
