### 2019-03-09 #1244 @ScottKolo
Yes. We removed it for most as you suggested, since it wasn’t helpful in those contexts, but left it when we referenced its deployment location.
