### 2019-03-10 #1244 @arfon
@ScottKolo - the paper currently looks a little peculiar with both the figures being right at the start. This PR moves them into the body of the paper: https://github.com/ScottKolo/suitesparse-matrix-collection-website/pull/7
